refactor: extract accounting readonly routes
